Let digits toggle axes nav only if they correspond to an existing axes.
Currently, pressing a (1-9) digit key while the cursor is on an axes
will change whether axes, in general, are responsive to "navigation" UI
(pan/zoom/interactive cursor): if the cursor is over axes *m* and digit
*n* is pressed, then axes *m* will become unresponsive to navigation UI
if m!=n, and become responsive again if m==n.

That's likely not the originally desired behavior: in the patch below, I
think the check `a.in_axes(event)` (which has a FIXME comment...) should
be removed...

While this is getting sorted out, in the meantime, this patch changes
the behavior above to *not* disable axes *m* if m!=n and *n* is greater
than the total number of axes (i.e., in a case where *n* does not
correspond to any valid axes).  The use case here is to use "4" and "6"
as navigation keys for an interactive image stack viewer (unfortunately
one can't by default use either left/right or h/l as Matplotlib already
assigns default meanings for these keys).