Deprecate \stackrel.
Mathtext's \stackrel{a}{b} behaves subtly differently from latex's
\stackrel{a}{b}: mathtext makes `a` and `b` equal-sized whereas latex
makes `a` smaller than `b` (`a` is an "annotation" over `b`).  Given
that we can already stack expressions using the (admittedly less
practical, but standard) \genfrac, just deprecate \stackrel instead of
maintaining our own latex-like dialect.

Also undeprecate passing obj_type to the cbook.deprecated decorator(!),
given that this PR has a good reason to use it... and while we're at it,
reorder the kwargs docs for cbook.deprecated and cbook.warn_deprecated
to match the order in the signature.