Use _warn_external for deprecations warnings.
We'll probably switch to _warn_external everywhere at some point, but in
the meantime I think deprecations warnings are ones of those that
benefit the most from warn_external ("oh, that's the place that needs to
be fixed.").

Note the internal import to workaround the circular import loop between
cbook and cbook.deprecation.