Fix Flash of Unstyled Content
This also removes all remaining Flipcause integration code.

The problem is the <iframe> in the footer that is part of the Flipcause
integration. Removing it gets rid of the FOUC.

It appears that the entire Flipcause integration is not used anymore
since commit 5c3b3eff5f04690eb6c32e1907ada26e09c65d22, in which
the JavaScript Flipcause buttons were replaced with normal links. That
commit removed the JavaScript part of the Flipcause integration, but
not the styles.