Give Carreau the ability to mention the backport bot.
Mostly because when the backport bot crashes and pull-requests get not
backported he needs to fix the bot and annoy the core dev to re-mention
the bot.

Give Carreau only the ability to mention the bot only for backport (no
other commands). That avoid having him to push a temporary patch on the
bot itself.