Clarify documentation for textprops keyword parameter of TextArea
As described in #12121, the current documentation of `textprops` keyword
parameter strongly, and erroneously, suggests that `textprops` should be a
FontProperties object. In reality, it should instead be a dictionary of keyword
parameters that will be passed to the `Text` object inside of `TextArea`. This
commit changes the documentation to reflect the current behavior.

closes #12121