Deprecate passing drawstyle with linestyle as single string.
Instead of `plt.plot(..., linestyle="steps--")`, use `plt.plot(...,
linestyle="--", drawstyle="steps")`. `ds` is now an alias for
`drawstyle`.

Note that `plt.plot(..., "steps--")` already doesn't work (due to
confusion as to whether "steps" is a color, a marker or a drawstyle).

The end goal is to simplify and rationalize linestyle specification as
rcParams vs. as arguments.