Improve formatting of imshow() cursor data when a colorbar exists.
When a colorbar exists, use its formatter to format cursor data.
For example, after
```
imshow([[10000, 10001]]); colorbar()
```
currently the cursor data on either pixel is rendered as 1e4, but after
this patch it becomes 0.0+1e4 / 1.0+1e4, or 10000.0 / 10001.0 if
`rcParams["axes.formatter.useoffset"]` is set to False.  (Even though
the version with the offset text may not be the most esthetic, it's
clearly more informative than the current behavior...)

It would be nice if this worked even for ScalarMappables that don't
have a colorbar; this may include extracting the Formatter selection
code out of the colorbar code into something generally applicable to
ScalarMappables, or just generating a hidden colorbar "on-the-fly" if
needed just for the purpose of getting its Formatter.