Prepare for ipython 7.1 stricter ipython directive.
In 7.1 the IPython directive will be stricter about unexpected
error/warnings _by default_ and will stop the build.

This can be disabled globally with ipython_warn_is_error,
or on a per-block basis by setting the appropriate `:okwarning:`/`okexcept`

I was only able to go that far as other things were preventing me from
finishing the build (some document not included in any toctree),
but at least this is one less error.