Tell IPython the correct GUI event loop to use for all backends.
IPython currently uses a hard-coded table
(IPython.core.pylabtools.backend2gui) to know which event loop to use
for which backend.  This approach fails for both the new builtin
cairo-based backends (which do not appear in the table), and for
third-party backends (e.g. mplcairo).

mplcairo has used some custom code to patch that table for a while;
reuse it to more generally use the new "required_interactive_framework"
attribute.

Note that this PR suggests that there should be a better way to go back
from the canvas class to the backend module (rather than looking for
`sys.modules[cls.__module__].required_interactive_framework`, which is a
bit hacky and could in theory fail if the canvas class is actually
defined in another submodule).