Allow "real" LaTeX code for pgf.preamble in matplotlibrc
This allows to use the rc key "pgf.preamble" to contain "real" LaTeX
code. As of before this patch, commas are used as parser separators,
causing the PGF backend call to latex to fail, for example at loading
packages which contain options such as:
\usepackage[protrusion=true, expansion=false]{microtype}

Passing a list of strings from within a Python script works fine, and
with this patch it stays this way.