Delete some `if __name__ == "__main__"` clauses.
In offsetbox and mpl_axes, they are just examples (mpl_axes doesn't even
call `show()`); if anything the examples belong in the examples/
directory but they don't look particularly worth moving there either.

In rcsetup, the clause is checking that the defaults in rcParams don't
get modified by the validators.  If anything this should be a test, but
it doesn't look like it mattered.  I did fix `text.latex.preamble`,
`pgf.preamble`, and `keymap.fullscreen` to be idempotent under the
validator; now the only rcParam that gets changed is `savefig.bbox` (due
to the presence of the "standard" / None synonym) but again this isn't a
problem.