BUG: Ensure that distinct polygon collections are shaded identically
Right now, shading ratios are computed by computing a shade in [-1, 1] for each face from the normal vector, and then normalizing across the polygon collection so that the brightest face has shade=1, and the dimmest shade=0.5.

This behaves poorly with multiple polygon collections, as the normalization is internal to each.
This results in two polygons with the same orientation but in different collections ending up with different shadings.
Notably, this obstructs the shading of `ax.voxels()`.

This approach instead scales `[-1, 1]` to `[0.4, 1]`, ignoring the local brightness minima and maxima within a collection.

Some more noticeable effects of this change:
* Highlights will be darker, unless they were already directly facing the light source
* Contrast between extreme highlights and shadows will be slightly greater (from 1:0.5 to 1:0.4)