Simplify stride calculations in loglocator.
These are just algebraic manipulations:

In classic mode, we want the smallest stride so that

    numdec/stride+1 <= numticks

i.e.

    stride >= numdec/(numticks-1)

so stride is the ceil() of the RHS (and at least 1, to handle the case
vmin==vmax).

In nonclassic mode, we want the smallest stride so that

    numdec//stride+1 <= numticks, i.e. numdec//stride < numticks

(everything is integer), i.e.

    numdec < numticks*stride- 1

i.e.

    stride > (numdec+1)/numticks

so stride is 1 + the floor() of the LHS.