Simplify use of Path._fast_from_codes_and_verts. (#12868)
All uses of Path._fast_from_codes_and_verts create the internals dict by
copying the attribute from another Path.  Instead of having the caller
create that dict, just let him pass a Path from which to copy the
attributes.

Note that this means that the copying will now also use the
simplify_threshold from the previous Path; however I think it is clear
that the previous behavior of not copying it was an oversight
(compounded by the fact that simplify_threshold is rarely, if ever,
overwritten on a per-Path basis).