Path: change the type of the code type constants to match the type of the codes array
The matching types make comparisons between the constants and values of the codes array faster.

Using the Gtk3Cairo backend with wire3d_animation_sgskip.py:
Before: 9.95 fps
After: 15.26 fps

The main areas where this helps for the cairo case is the faster comparisons in
Path.iter_segments() and in _append_paths() of the cairo backend.