Make GridSpec.update docstring match behavior.
The sentence "If any kwarg is None, default to the current value, if
set, otherwise to rc." is clearly wrong given the lines just below,
which overwrite the current value unconditonally:

    for k, v in kwargs.items():
        if k in self._AllowedKeys:
            setattr(self, k, v)

This can also be checked with e.g.

    from matplotlib.gridspec import GridSpec
    from matplotlib import pyplot as plt
    gs = GridSpec(1, 1, left=.5)
    gs.update(left=None)
    plt.figure().add_subplot(gs[0, 0])
    plt.show()

where the update reset left to the rc value.

Given that the behavior has been like that ever since gridspec was added
to matplotlib, it seems better to update the docstring.