Deprecate case-insensitive colors.
The motivation is explained in the changelog entry.

Note that even though

    plt.plot([1, 2], [3, 4], "Y")

"worked" so far,

    plt.plot([1, 2], [3, 4], "Yo")

(for example), i.e. additionally passing a marker spec, has not worked
at least since matplotlib 1.5 (a ValueError is thrown).  As this does
not appear to have ever been raised on the issue tracker, I believe the
number of users of UPPERCASE colors must be limited.