Defer checking of tex install to when it is actually used.
The call to checkdep_usetex() in `__init__.py` doesn't help for people
explicitly passing usetex=True.  Moreover, it slows down import time
for those who *do* have text.usetex set to True in their rcfile by ~20ms
(when timing this, note that checkdep_ghostscript (which checkdep_usetex
calls) has a cache, which should be disabled for timing purposes.

Moreover, checkdep_usetex is overly broad, always requiring dvipng
(which is not required for pdf/ps output) and ghostscript (which is not
required for raster output).

Instead, move the check to texmanager.py, which will provide a more
meaningful error message to end users, speed up import time (20ms is
similar to the import time of pyparsing or urllib.request, for example)
and avoid the inaccuracies above.