Avoid unneeded copies from flatten().
flatten() always returns a new copy, but nearly all places that use
flatten() in the codebase don't need to pay that cost; the only place
that needed it did an additional copy.  Fix them.

As a replacement, there's ravel(), which only makes a copy if the input
is not C-contiguous to start with, and reshape(..., -1) which never
makes a copy (as long as the input is already an ndarray), so prefer the
latter.

Note that in the changed snippet in mplot3d, the AttributeError would
never be raised, as x/y/z are converted to arrays and thus always have a
flatten() method.