Remove/rework uses of np.where where possible.
In a few places, np.where can be replaced by direct use of boolean
arrays, or other constructs.

In colors.py, the use of putmask may be slightly obscure (so I left a
comment explaining it), but avoids repeating the modified array twice,
so I think is more legible.

In polar.py, preallocating xy is just pointless given that we're
constructing intermediate arrays for x and y and filling them into xy
afterwards.