Add nonsingular to the locator base class, and use it in set_*lim too.
Currently, some Locator subclasses define a nonsingular() method, which
is used by autoscale_view() in case the autoscaling would return a
degenerate interval, to expand the view limits; autoscale_view() falls
back on a default in case the nonsingular() method does not exist.  Move
that default to the Locator base class.

Also use that nonsingular() method when a degenerate interval is passed
to set_xlim/set_ylim/etc., instead of always hardcoding the same linear
expansion.

(Semantically the correct place for this nonsingular() method is
probably on the scale class rather than on the locator...)

The changes to test_ticker are due to the fact that the default
expansion is now by 5%, so there's no offset text anymore in the
left=right=123 case (as it gets expanded to 116.85, 129.15).