Add get/set_in_layout to artist API docs.
They don't seem to fit in any existing subsection and creating a section
just for them seems silly, so move them, together with other
single-entry subsections, to "misc.".

To compensate the loss of the "margins and autoscaling subsection", add
mention of autoscaling in the first line of the sticky_edges docstring,
to keep that term ctrl-f'able on the API docs page.