Let boxplot() defer rcParams application to bxp()
instead of duplicating the defaults-applying code in both methods.

This caught an oversight in the application of the
boxplot.boxprops.linewidth rcParam in bxp (see changelog note);
because of that, the bxp testing code needed some update (to keep the
old behavior).

Preliminary work to having bxp() normalize properties passed in the
boxprops, meanprops, etc. kwargs as well.