Deprecate TextWithDash.
TextWithDash appears to be an ancestral version of Annotation that was
added over two commits in 2005 (6a4d343, fe80816), revised once in 2006
and another time in 2008, and never touched since then.

While it "works", its implementation is 400 lines long, it is only
mentioned by a single example (dashpointlabel.py), somewhat complicates
Axes.text(), and is likely rarely used and replaceable by Annotation for
most use cases.

Hence, deprecate it, keeping open the option to undeprecate if this
turns out to be too disruptive.