Deprecate MovieWriterRegistry cache-dirtyness system.
MovieWriterRegistry has a fairly extensive system to cache whether movie
writer classes are available or not.  Deprecate it and always perform
the check every time: for nearly all classes, the availability check is
either an import check (Pillow), or checking for whether an executable
is in the $PATH -- this is negligible compared to actually running the
executable to encode the movie.

The only exception is that ffmpeg actually checks whether it is
actually ffmpeg or ubuntu's broken libav.  In order to avoid performing
that check (... which should still be not so slow) when Pillow is
available (Pillow is registered first, so comes first :)), add
`MovieWriterRegistry.__iter__` which allows to get the first available
writer without checking the availability of the subsequent ones (this
behavior is consistent with `MovieWriterRegistry.__getitem__` and
`MovieWriterRegistry.list`).