Clarify behavior of the 'tight' kwarg to autoscale/autoscale_view.
The 'tight' kwarg behaves subtly differently in autoscale and in
autoscale_view (in the former, it additionally sets the margins to
zero).

In autoscale_view, the default of tight=True actually has no effect in
the default style, which disables it via an rcParam (which, by the way,
is documented with a typo in matplotlibrc.template).

Try to clarify these less-than-obvious behaviors.