FIX: if the user calls `matplotlib.use` disable backend fallback
During pyplot import if `rcParmas['backend_fallback'] == True` and we
find the user has requested an inconsistent backend with the currently
running event loop we reset `rcParams['backend']` to the auto backend
selection sentinel.  Immediately afterwards we call `switch_backend`
which will then select a backend consistent with the currently running
event loop.

In the case where we are on a headless linux machine we always detect
the running event loop as 'headless'.  This means that if any
interactive backend was selected (via an rcparams file, directly
setting `rcParams['backend']` or using `mpl.use`) on import of
`pyplot` we would, by default, run through the fallback logic and
select 'agg' as the backend.

This change records the user's intent to select a backend (via `use`,
the MPLBACKEND environmental variable, or directly setting
`rcParams['backend']`) by setting `rcParam['backend_fallback'] =
False` to prevent this fallback in the case where the user has
explicitly selected a backend in their code before pyplot is imported.

If the backend is selected an rcparam file the value of
`'backend_fallback'` in the file will be respected (and defaults to True).

closes #13883