FIX: do not try to manage the visibility of un-drawn ticks
This reverts part of the changes from #13908 (in particular commit
e908efe37131da0697bd8e9c62444d37b4940ad3 ) that tried to manage the
visibility if the ticks that are not used.  However, there was a bug
in this that nothing would ever set ticks to be visible
again.  Because we currently cache / recycle ticks this meant that
panning / zooming or resizing the figure would eventually lose ticks
if the number of ticks ever went down and then back up.

We could fix this by adding a `tick.set_visible(True)` for the drawn
ticks, however this would break the (brittle) use case where users are
reaching in and manually setting the visibility of individual ticks to
False to hide them.

Instead, we revert the consistency code altogether and accept that
there will be `Tick` objects in `Axis.majorTicks` and
`Axis.minorTicks` that are marked as visible, but are not
drawn (because they are filtered by `_update_ticks`).

Closes #14054