Clarify error with usetex when cm-super is not installed.
This PR makes the error when cm-super is not installed (and
rcParams["text.latex.unicode"] = True, which is now the default) be
```
! LaTeX Error: File `type1ec.sty' not found.

Type X to quit or <RETURN> to proceed,
or enter new name. (Default extension: sty)

Enter file name:
! Emergency stop.
<read *>

l.5 \usepackage
               {type1ec}^^M
No pages of output.
```
(or, if using miktex, miktex will prompt for installing cm-super).