Isolate nbagg test from user ipython profile.
My ~/.ipython profile causes test_backend_nbagg to fail (but is
otherwise valid for my own use).  To fix that, point the ipython
directory to a temporary directory when running that test (per
https://ipython.readthedocs.io/en/stable/config/intro.html#the-ipython-directory).

While we're at it, also create the temporary output file in that
directory -- that's quite likely to be necessary for the test to
actually work on Windows, anyways (due to vagaries with writing from
multiple processes to a NamedTemporaryFile).