Assume that mpl-data is in its standard location.
The previous candidates tested in get_candidate_paths were for

1) when matplotlib itself was a namespace package (a long time ago,
   before mpl_toolkits were move out of matplotlib proper), and
2) py2exe support, which is deprecated (see deprecation of
   get_py2exe_datafiles).

Just to be sure in case someone else is relying on the paths in
get_candidate_paths, emit a deprecation warning in case they are needed,
but let's not bother with an API changes note (it's not even clear how
to word that).

Note that redistributors (e.g. linux distro packagers) who may want to
move mpl-data out can still do so by patching out the entire
`_get_data_path` function.