Simplify implementation of svg.image_inline.
RendererSVG._imaged is basically a single-entry dict (because
Renderer.basename never changes), so just replace it by a counter.

Also replace the assert by a proper exception, and only trigger it if
the user actually tries to save an image while the filesystem path is
unknown (otherwise, there's no problem).