Deprecate checkdep_usetex.
See changelog for rationale.  Note that checkdep_usetex was *only*
used to create a test decorator.

This also goes well with the deprecation of all other checkdep_foo
helpers in the matplotlib toplevel module.

Instead, just directly write the proper test decorator.  Use
unittest.skipIf instead of pytest.skipif as pytest handles the unittest
decorator just as well, and this makes the decorator work if someone
*really* wants to use it in a non-pytest test suite.