Fix uses of PyObject_IsTrue.
PyObject_IsTrue can return -1 if an exception occurred when trying to
convert a Python object to a bool (a typical case is for numpy arrays,
which cannot be converted to bools).  Failure to handle this correctly
results e.g. in
```
l, = plt.plot([1, 2]); l.get_path().should_simplify = np.array([1, 2])
```
```
ValueError: The truth value of an array with more than one element is ambiguous. Use a.any() or a.all()

The above exception was the direct cause of the following exception:

Traceback (most recent call last):
  File "/usr/lib/python3.7/site-packages/matplotlib/backends/backend_qt5.py", line 501, in _draw_idle
<elided>
  File "/usr/lib/python3.7/site-packages/matplotlib/backends/backend_agg.py", line 146, in draw_path
    self._renderer.draw_path(gc, path, transform, rgbFace)
SystemError: PyEval_EvalFrameEx returned a result with an error set
```
whereas after this PR, the ValueError gets correctly propagated out.

No tests, because one really needs to go out of their way to trigger the
faulty cases anyways...