Simplify Qt tests.
In test_backend_qt, qt_key is basically a complicated way to access Qt
modifier keys using nonstandard names that are not used anywhere but in
the tests (ControlKey instead of Key_Control).  Get rid of it and just
use standard Qt modifier key names.  Simplify the qt_module fixture as a
consequence (just returning QtCore).