TST: Change tests to use default styles
As requested in review feedback - since all the images changed, may as well
restyle them.

Note that this restyling caused a test to become platform-dependent on the
anti-aliasing on the edges of a narrow line - presumably due to the resolution
decrease making the line narrower - so a non-zero `tol` value was added.