Update some mplot3d docs.
Mostly removing "This function was added, but not tested" when codecov
indicates that the function is actually covered by CI (it may still not
be *directly* tested, but that's the case of a bunch of APIs and we
don't bother with peppering their docstrings with such statements).

The get_zbound docstring is adapted from the one of get_xbound.