Soft-deprecate transform_point.
The transform() method has explicit code to handle 1d input, so
transform_point() is redundant with transform() (... and is anyways
implemented in terms of transform(), so not faster either)... and longer
to type.

Replace it throughout the codebase.  Also compress consecutive calls
to transform_point() with a single call to transform(), which may be a
bit faster (one matrix multiplication instead of two); also a single
replacement to use transform_bbox as appropriate.

Only "deprecate" transform_point in the docs because I don't really want
to discuss here whether this would be too disruptive...