[Experiement] Auto-expand signature when **kwarg passed to other functions.
Many functions current takes **kwargs and pass them down. During a SciPy
discussion it was brought up that this was sub-optimal for new user.

We can use the fact that signature can be faked, and inspect to
propagate the actual keyword arguments and substitute them for **kwargs
in another.

With this, the signature of pyplot.subplots(), which currently as viewed
in IPython is

    plt.subplots(
        nrows=1,
        ncols=1,
        sharex=False,
        sharey=False,
        squeeze=True,
        subplot_kw=None,
        gridspec_kw=None,
        **fig_kw,
    )

now becomes:

    plt.subplots(
        nrows=1,
        ncols=1,
        sharex=False,
        sharey=False,
        squeeze=True,
        subplot_kw=None,
        gridspec_kw=None,
    +   num=None,
    +   figsize=None,
    +   dpi=None,
    +   facecolor=None,
    +   edgecolor=None,
    +   frameon=True,
    +   FigureClass=<class 'matplotlib.figure.Figure'>,
    +   clear=False,
    +   **kwargs, # used to be **fig_kw, to fix.
    )

... which is not totally correct (they are kwargs only...) but you
should see the point.

This is nice as:

1) it now appear in IPython ? and ??
2) tab completion pick up these kwargs.

And it should have minimal overhead at import time, and is so far
limited to function that pass all of **kwarg.

Many places in matplotlib are also too dynamic to properly do it, but
could allow a manual variant.