Inline some afm parsing code.
Splitting _parse_afm out of `__init__` doesn't buy much, and its
docstring is not particularly informative either, so just inline
it to the AFM constructor.  Likewise, most of the benefits from
`_sanity_check` can be obtained by just looking for the first entry of
the header (this admittedly relies on dict-order-maintaining, which is
present on Py3.6 though not mandated (it is mandated in Py3.7), but I'm
not aware of Matplotlib running on an alternative Py3.6 that does not
support that behavior...).