Get default params from matplotlibrc.template.
This makes matplotlibrc.template a valid style file representing the
default style (modulo things that styles do not control, such as the
backend), avoids having to duplicate the defaults in rcsetup, removes
the need to test that the default matplotlibrc is up to date (if not,
matplotlib fails to import), and makes it easier to fit lines within 79
characters in rcsetup.py.

The only tricky points are:
- datapath is not really an rcparam anyways, so just hack `__getitem__`
  to always return the correct value.
- the entry for path.effects was incorrect, as [] would be validated as
  the "[]" list which is not a valid list of patheffects.  In fact
  this rcParam cannot be meaningfully set from matplotlibrc; one would
  need to do an eval() like for axes.prop_cycle but let's not get
  there...  I changed the validator to validate_anylist, which at least
  works for the default of an empty list...
- we need to be a bit more careful when constructing the global rcParams
  instance as well as rcParamsOrig, rcParamsDefault, to not resolve
  _auto_backend_sentinel too early.

One can check that the rcParams are unchanged by printing them out --
that catches a typo: two entries in font.fantasy should be "Impact",
"Western"; not "ImpactWestern".