Simplify Tick and Axis initialization.
Instead of delegating to a bunch of private methods (_get_text1,
_get_text2, _get_tick1line, _get_tick2line, _get_gridline, _get_label,
_get_offset_text) to construct the sub-artists (lines, labels, etc.) of
Ticks and Axises, just set up some default instances in the base class
constructor and make additional customizations in the subclass
constructor.

A quick grepping suggests (no guarantees) that neither cartopy nor
wcsaxes override these private methods (an API where third-party
subclasses are encouraged to override private methods is a bit
awkward...) and in any case the approach of customizing the sub-artists
in the subclass' `__init__` works with all versions of Matplotlib.