Use fix_minus in format_data_short.
See changelog.  AFAICS, all GUI toolkits (at least qt5/gtk3/wx/tk)
display the unicode minus properly.

Moving the implementation of the replacement to the base Formatter class
is because it is otherwise confusing as to whether `self.fix_minus`
does the replacement or not -- one needs to check whether the current
class overrides the do-nothing fix_minus.  Instead, one can just have a
base-class implementation that performs the replacement, and not call
`self.fix_minus` if the replacement is not desired.

For example, LogFormatter.fix_minus used to do nothing; there are tests
that check that.  Just don't call fix_minus instead.