FIX: Don't enable IPython integration if not entering REPL.
IPython can also be used to simply run files; when this is the case we
are not going to be interactive; and we do not want to enable the
eventloop integration. Otherwise someone doing:

    $ ipython script.py

Will simply see a window blink, as IPython will exit immediately instead
of showing the window and exit.

I'm not quite sure how to test that are we basically _want_ something to
block forever and wait for user input; so no test added

Closes: https://github.com/ipython/ipython/issues/11837