Ensure streamplot Euler step is always called when going out of bounds. (#11947)
* FIX: Ensure Euler step is always called when going out of bounds.

This makes figures neater. Also fix places that were using
dmap.grid.nx for the width of the grid, which should be
dmap.grid.nx - 1.

* Remove unused code

* Improve the OutOfBounds logic in the integration.

Moves the raising to the gradient steps simplifying the integrator code a bit. Also makes OutOfBounds inherit from IndexError as suggested by tacaswell.

* TST: push up the tolerance on mac

We are seeing 1bit differences on 2 pixel for this test between
linux/win vs mac.