MNT: remove the swap_vlim dance
In https://github.com/matplotlib/matplotlib/pull/14624 we reverted
changing the order of the limits to match the input order.  Doing the
same here for consistency.

Sort the limits before comparing.  We want to make sure that it works
with the values reversed, but do not expect non-singular to preserve
the order.