Embedding in Tk example: Fix toolbar being clipped.
Calling `pack` without options puts the widget at the end of the package
list. Space is allocated to widgets sequentially. If a window is too
small, there may be no space left for widgets. By packing the canvas
last, we essentially give the toolbar and the quit button priority in
the packing order.

Fixes #13485.