Sphinx extension: support captions in inline plots.
This commit adds a :caption: option to the plot directive provided by the
Sphinx extension (matplotlib.sphinxext.plot_directive). Without this
option, there is no way to specify a caption for a plot generated from
inline content.

This is fully backwards-compatible. If a plot directive with a path to a
source file has both a :caption: option and content provided, the
content is used for the caption and the option is ignored.